Add library module
